Moved helper methods to parse
